Add tsx monthpicker
