+++ .expo/packager-info.json
@@ -1,8 +1,8 @@
 {
   "devToolsPort": 19002,
   "expoServerPort": 19000,
-  "packagerPort": 19001,
-  "packagerPid": 8872,
+  "packagerPort": 19003,
+  "packagerPid": 72525,
   "webpackServerPort": 19006,
   "expoServerNgrokUrl": null,
   "packagerNgrokUrl": null,
