+++ package.json
@@ -15,6 +15,7 @@
     "@react-navigation/material-bottom-tabs": "^5.3.0",
     "@react-navigation/native": "^5.8.0",
     "@react-navigation/stack": "^5.10.0",
+    "@types/react-router": "^5.1.8",
     "date-fns": "^2.16.1",
     "dotenv": "^8.2.0",
     "expo": "^39.0.3",
@@ -35,7 +36,9 @@
     "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
-    "@babel/core": "~7.9.0"
+    "@babel/core": "~7.9.0",
+    "@types/react": "^16.9.55",
+    "@types/react-native": "^0.63.32"
   },
   "private": true,
   "name": "SaveDatMoney"
